ml/requirements-ml.txt
torch
torchvision
Pillow
numpy
tqdm
